[feat] : 회원 잔여 포인트가 남아있는 경우 예외처리 추가

+++ src/main/java/com/shoekream/service/UserService.java
@@ -95,6 +95,10 @@ public class UserService {
 
         foundUser.checkPassword(encoder, request.getPassword());
 
+        if (foundUser.hasPoint()) {
+            throw new ShoeKreamException(WITHDRAWAL_NOT_ALLOWED_POINT);
+        }
+
         userRepository.delete(foundUser);
 
         return foundUser.toUserResponse();
